Move global type Deal to the feed-parser
